+++ src/main/java/ru/mail/polis/service/shkalev/ServiceUtils.java
@@ -3,6 +3,7 @@ package ru.mail.polis.service.shkalev;
 import one.nio.http.Request;
 import one.nio.http.Response;
 import org.jetbrains.annotations.NotNull;
+import ru.mail.polis.dao.shkalev.Row;
 
 final class ServiceUtils {
     static final String PROXY_HEADER = "Is-Proxy: True";
@@ -22,6 +23,19 @@ final class ServiceUtils {
 
     static boolean validResponse(@NotNull final Response response) {
         return response.getHeaders()[0].equals(Response.NOT_FOUND) || response.getHeaders()[0].equals(Response.OK);
+    }
+
+    static Response responseFromRow(@NotNull final Row row){
+        if (row.isDead()) {
+            final Response response = new Response(Response.NOT_FOUND, Response.EMPTY);
+            response.addHeader(ServiceUtils.TIME_HEADER + row.getTime());
+            return response;
+        }
+        final byte[] body = new byte[row.getValue().remaining()];
+        row.getValue().get(body);
+        final Response response = new Response(Response.OK, body);
+        response.addHeader(ServiceUtils.TIME_HEADER + row.getTime());
+        return response;
     }
 
 }
